- Refactor: Move all topsystem related methods to separate subclass. - Fix bug where GHB package had drainage aggregation methods - Remove tests for non-topsystem packages.
