+++ imod/mf6/drn.py
@@ -9,11 +9,11 @@ from imod.common.utilities.dataclass_type import DataclassType
 from imod.common.utilities.mask import broadcast_and_mask_arrays
 from imod.logging import init_log_decorator, standard_log_decorator
 from imod.mf6.aggregate.aggregate_schemes import DrainageAggregationMethod
-from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.disv import VerticesDiscretization
 from imod.mf6.npf import NodePropertyFlow
 from imod.mf6.regrid.regrid_schemes import DrainageRegridMethod
+from imod.mf6.topsystem import TopSystemBoundaryCondition
 from imod.mf6.utilities.imod5_converter import regrid_imod5_pkg_data
 from imod.mf6.utilities.package import set_repeat_stress_if_available
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
@@ -40,7 +40,7 @@ from imod.typing.grid import enforce_dim_order, has_negative_layer, is_planar_gr
 from imod.util.regrid import RegridderWeightsCache
 
 
-class Drainage(BoundaryCondition, IRegridPackage):
+class Drainage(TopSystemBoundaryCondition, IRegridPackage):
     """
     The Drain package is used to simulate head-dependent flux boundaries.
     https://water.usgs.gov/ogw/modflow/mf6io.pdf#page=67
@@ -132,7 +132,7 @@ class Drainage(BoundaryCondition, IRegridPackage):
 
     _period_data = ("elevation", "conductance")
     _keyword_map = {}
-    _template = BoundaryCondition._initialize_template(_pkg_id)
+    _template = TopSystemBoundaryCondition._initialize_template(_pkg_id)
     _auxiliary_data = {"concentration": "species"}
     _regrid_method = DrainageRegridMethod()
     _aggregate_method: DataclassType = DrainageAggregationMethod()
@@ -187,7 +187,7 @@ class Drainage(BoundaryCondition, IRegridPackage):
         super().__init__(cleaned_dict)
 
     @classmethod
-    def allocate_and_distribute_planar_data(
+    def _allocate_and_distribute_planar_data(
         cls,
         planar_data: dict[str, GridDataArray],
         dis: StructuredDiscretization | VerticesDiscretization,
@@ -328,7 +328,7 @@ class Drainage(BoundaryCondition, IRegridPackage):
         regridded_package_data = broadcast_and_mask_arrays(regridded_package_data)
         is_planar = is_planar_grid(regridded_package_data["elevation"])
         if is_planar:
-            layered_data = cls.allocate_and_distribute_planar_data(
+            layered_data = cls._allocate_and_distribute_planar_data(
                 regridded_package_data,
                 target_dis,
                 target_npf,
